blkin: make encoding cross-version safe
We can't have divergent encodings based on compilation options.
Instead, increment HEAD_VERSION for MOSD[Sub]Op[Reply] and encode
tracing data in a bufferlist. This makes it simple to detect when
no trace data is there, or to skip it on an osd compiled without blkin.

This lets osds with and without blkin talk to each other,
with little overhead for OSDs without blkin.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>